Move `ringbuffer` to internal
