+++ cmd/ringdump/ringdump.go
@@ -4,9 +4,10 @@ import (
 	"bytes"
 	"flag"
 	"fmt"
-	"github.com/usnistgov/dastard/packets"
-	"github.com/usnistgov/dastard/ringbuffer"
 	"time"
+
+	"github.com/usnistgov/dastard/internal/ringbuffer"
+	"github.com/usnistgov/dastard/packets"
 )
 
 func dumpdata(data []byte, max int) {
